refactor: 멘토님 피드백 반영 (#235)
