+++ src/pages/Like.tsx
@@ -8,6 +8,7 @@ import { useParams } from 'react-router-dom';
 import throttle from 'lodash/throttle';
 import EmptyMessage from '@/components/EmptyMessage';
 import Loading from '@/components/Loading';
+import useFetchUserInformation from '@/hooks/useFetchUserInformation';
 
 interface LikedPlaylistData {
   title: string;
@@ -21,43 +22,13 @@ interface LikedPlaylistData {
   profileImage: string;
 }
 
-interface UserInformation {
-  profileImage: string;
-  userName: string;
-  userId: string;
-}
-
 const Like: React.FC = () => {
   const [visibleItems, setVisibleItems] = useState(8);
   const [loading, setLoading] = useState(false);
   const [likedPlaylists, setLikedPlaylists] = useState<LikedPlaylistData[]>([]);
   const [error, setError] = useState<string | null>(null);
-  const [userInformation, setUserInformation] = useState<UserInformation | null>(null);
-
   const { userId } = useParams<{ userId: string }>();
-
-  const fetchUserInformation = async () => {
-    if (!userId) {
-      setError('유저 ID가 없습니다.');
-      return;
-    }
-
-    try {
-      const response = await fetch(`/api/profile/${userId}`);
-      if (!response.ok) {
-        throw new Error('사용자 정보를 가져오는 중 오류가 발생했습니다.');
-      }
-      const data = await response.json();
-      setUserInformation(data);
-    } catch (e) {
-      console.error('사용자 정보 요청 오류:', e);
-      setError('사용자 정보를 불러오는 중 오류가 발생했습니다.');
-    }
-  };
-
-  useEffect(() => {
-    fetchUserInformation();
-  }, [userId]);
+  const { userInformation } = useFetchUserInformation(userId);
 
   useEffect(() => {
     const fetchLikedPlaylists = async () => {
